% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqtrunc.R
\name{qqtrunc}
\alias{qqtrunc}
\title{Create a q-q plot for a truncated random variable.}
\usage{
qqtrunc(x, spec, a = -Inf, b = Inf,
  title = "Truncated Distribution Q-Q Plot",
  xlabel = "Theoretical Quantiles", ylabel = "Sample Quantiles",
  params = NULL, ...)
}
\arguments{
\item{x}{A numeric vector of sample values}

\item{spec}{A character vector for the name of the distribution (e.g., "norm")}

\item{a}{The minimum value allowed in the distribution (inclusive).}

\item{b}{The maximum value allowed in the distribution (inclusive).}

\item{title}{Plot title}

\item{xlabel}{Label for the X-axis}

\item{ylabel}{Label for the Y-axis}

\item{params}{A list of distribution parameter name/value pairs}

\item{...}{Distribution parameter arguments. Only used if `params` is NULL.}
}
\description{
This function produces a q-q plot of sample quantiles against theoretical
quantiles for a truncated random variable
}
